将 ServerStopEvent 移到单独的类
